enhance examples
