+++ docs/guides/cookbook/overview.mdx
@@ -20,6 +20,8 @@ Here's a list of common examples of how to use Latitude. The list will grow with
 - [Creating a prompt](https://github.com/latitude-dev/latitude-llm/blob/main/examples/sdks/python/create_prompt.py)
 - [Running a prompt](https://github.com/latitude-dev/latitude-llm/blob/main/examples/sdks/python/run_prompt.py)
 - [Running a prompt with tools](https://github.com/latitude-dev/latitude-llm/blob/main/examples/sdks/python/run_prompt.py)
+- [Running a prompt locally with OpenAI](https://github.com/latitude-dev/latitude-llm/blob/main/examples/sdks/python/providers/with_openai.py)
+- [Running a prompt locally with Anthropic](https://github.com/latitude-dev/latitude-llm/blob/main/examples/sdks/python/providers/with_anthropic.py)
 - [Chaining multiple calls with a growing context (chatbot)](https://github.com/latitude-dev/latitude-llm/blob/main/examples/sdks/python/run_prompt.py)
 - [Pushing external logs to Latitude for evaluation](https://github.com/latitude-dev/latitude-llm/blob/main/examples/sdks/python/upload_logs.py)
 - [Triggering LLM as judge evaluations](https://github.com/latitude-dev/latitude-llm/blob/main/examples/sdks/python/evaluate_chat.py)
